feat: render NFT data from user wallet
