+++ packages/db/src/schema/index.ts
@@ -8,4 +8,3 @@ export * from "./Member";
 export * from "./Status";
 export * from "./TwoFacCode";
 export * from "./Tag";
-
